+++ changelog.md
@@ -1,11 +1,11 @@
 # Changelog
 
-* Config option for damaging enchantments
 * Blast furnace multi ghost blocks
 
 ## Upcoming
 * Protection enchantment is now treasure
 * Absorption is now affected only by damage types outside the minecraft:bypasses_armor damage type tag
+* Added a config option to disable the replacing of damaging enchantments
 * Further reduced Shovels base damage (3 -> 2.5)
 * Fixed 'Player's killer bounty' name changing before the player's death
 * Fixed diamond and netherite armor having vanilla toughness